lyceum/requirements/dev.txt
-r prod.txt
black==24.8.0
django-debug-toolbar==4.4.6
flake8==7.1.1
flake8-import-order==0.18.2
flake8-quotes==3.4.0
isort==5.13.2
pep8-naming==0.14.1